Unit tests for Places Data sources

+++ app/src/main/java/com/wks/nearby/data/places/source/PlacesDataSource.java
@@ -26,7 +26,7 @@ public interface PlacesDataSource {
 
     void loadNearbyPlaces(double latitude,
                           double longitude,
-                          double radius,
+                          long radius,
                           @NonNull LoadNearbyPlacesCallback callback);
 
     void loadPlaceDetails(@NonNull String placeId,
